+++ app/controllers/competitions_controller.rb
@@ -3,7 +3,8 @@ class CompetitionsController < ApplicationController
   include Contracts
   
   def competition_summary(competition)
-    Listtable.new(view_context).records(competition).columns([:name, :short_name, :competition_type, :city, :country, :site_url, :start_date, :end_date, :comment])
+    Listtable.new(view_context).records(competition).
+      columns([:name, :short_name, :competition_type, :city, :country, :site_url, :start_date, :end_date, :timezone, :comment])
   end
     
   def result_type(category, segment)
@@ -18,9 +19,8 @@ class CompetitionsController < ApplicationController
   def result_datatable(competition, category, segment)
     case result_type(category, segment)
     when :category
-      Datatable.new(view_context).records(competition.results.category(category).includes(:skater, :scores)).columns([:ranking, :skater_name, :nation, :points,
-                                                                                                                      :short_ranking, :short_tss, :short_tes, :short_pcs, :short_deductions, :short_bv,
-                                                                                                                      :free_ranking, :free_tss, :free_tes, :free_pcs, :free_deductions, :free_bv,]).tap {|d| d.default_orders([[:points, :desc], [:ranking, :asc]])}
+      Datatable.new(view_context).records(competition.results.category(category).includes(:skater, :scores)).
+        columns([:ranking, :skater_name, :nation, :points, :short_ranking, :short_tss, :short_tes, :short_pcs, :short_deductions, :short_bv, :free_ranking, :free_tss, :free_tes, :free_pcs, :free_deductions, :free_bv,]).tap {|d| d.default_orders([[:points, :desc], [:ranking, :asc]])}
       
     when :segment
       Datatable.new(view_context).records(competition.scores.category(category).segment(segment).order(:ranking).includes(:skater, :elements, :components)).
@@ -41,10 +41,7 @@ class CompetitionsController < ApplicationController
 
     respond_to do |format|
       format.html {
-        render :show, locals: data.merge(competition: competition,
-                                         category: category,
-                                         segment: segment,
-                                         )
+        render :show, locals: data.merge(competition: competition, category: category, segment: segment,)
       }
       format.json {
         render json: data