+++ app/controllers/concerns/error_handlers.rb
@@ -3,7 +3,7 @@ module ErrorHandlers
   
   included do
     unless Rails.env.development?
-    #f Rails.env.production?
+    #if false
       rescue_from Exception, with: :handler_500
       rescue_from ActiveRecord::RecordNotFound, with: :handler_404
       rescue_from ActionController::RoutingError, with: :handler_404