+++ app/parsers/competition_parser/isu_generic/summary_parser.rb
@@ -6,7 +6,7 @@ module CompetitionParser
       def parse(site_url, date_format:)
         page = get_url(site_url) || return
         city, country = parse_city_country(page)
-        #puts "#{city}, #{country}"
+
         competition = {
           name: parse_name(page),
           site_url: site_url,
@@ -29,7 +29,7 @@ module CompetitionParser
             competition[:segments][category][segment] = {
               panel_url: hash[:panel_url],
               score_url: hash[:score_url],
-              date: time_schedule.select {|item|
+              time: time_schedule.select {|item|      ## time ? date ? 
                 item[:category] == category && item[:segment] == segment
               }.first.try(:[], :time),
             }
@@ -37,6 +37,7 @@ module CompetitionParser
         end
         competition[:start_date] = time_schedule.map {|d| d[:time]}.min || Time.new(1970, 1, 1)
         competition[:end_date] = time_schedule.map {|d| d[:time]}.max || Time.new(1970, 1, 1)
+        competition[:timezone] = (time_schedule.present? ) ? time_schedule.first[:time].time_zone.name : "UTC"
 
         year, month = competition[:start_date].year, competition[:start_date].month
         year -= 1 if month <= 6
@@ -93,6 +94,15 @@ module CompetitionParser
         end
         summary
       end
+      ################
+      def get_timezone(page)
+        page.xpath("//*[contains(text(), 'Local Time')]").text() =~ / ([\+\-]\d\d:\d\d)/
+        local_tz = $1 || "+00:00"
+        $1 =~ /([\+\-]\d\d)/
+        utc_offset = $1.to_i
+
+        "Etc/GMT%+d" % [ utc_offset * -1]
+      end
       def get_time_schedule_rows(page)
         #page.xpath("//table[*[th[text()='Date']]]").xpath(".//tr")
         elem = page.xpath("//table//tr//*[text()='Date']").first || raise
@@ -103,10 +113,7 @@ module CompetitionParser
         rows = get_time_schedule_rows(page)
         dt_str = ""
         time_schedule = []
-        page.xpath("//*[contains(text(), 'Local Time')]").text() =~ / ([\+\-]\d\d:\d\d)/
-        local_tz = $1 || "+00:00"
-        $1 =~ /([\+\-])(\d\d)/
-        local_tz_hour = "%s%d" % [$1, $2.to_i]
+        timezone = get_timezone(page)
         
         rows.each do |row|
           next if row.xpath("td").blank?
@@ -119,14 +126,14 @@ module CompetitionParser
           tm_str = row.xpath("td[2]").text
           dt_tm_str = "#{dt_str} #{tm_str}"
           #dt_tm_str += " #{local_tz}" if tz == "UTC"
-          tz = "Etc/GMT#{local_tz_hour}"
+          #tz = "Etc/GMT#{local_tz_hour.to_i }"
 
           tm = 
             unless date_format.blank?
               Time.strptime("#{dt_tm_str}", "#{date_format} %H:%M:%S")
             else
               dt_tm_str
-            end.in_time_zone(tz)
+            end.in_time_zone(ActiveSupport::TimeZone[timezone])
           
           #next if tm.nil?
           tm = tm + 2000.years if tm.year < 100  ## for ondrei nepela
@@ -147,7 +154,7 @@ module CompetitionParser
       ###
       private
       def normalize_category(category)
-        category.squish.upcase.gsub(/^SENIOR /, '').gsub(/ SINGLE SKATING/, "").gsub(/ SKATING/, "")
+        category.squish.upcase.gsub(/^PAIR SKATING$/, "PAIRS").gsub(/^SENIOR /, '').gsub(/ SINGLE SKATING/, "").gsub(/ SKATING/, "")
       end
     end
   end