+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - date search
 - delete country-city mapping
 - panels: panel db entries
-- factory_bot
+- timezone
+- youtube register
+- team trophy
 
 ### new feature
 - embed youtube in score (or just link to search page ?)
@@ -14,6 +16,9 @@
 - app/controller/application_controller: unless Rails.env.develop
 - config/application: Version
 - config/competitions.yml: debug
+
+## 1.0.4-pre2
+- competition.date as Date
 
 ## 1.0.4-pre1
 - delete graphs on skater